Isolate image sequence creation from engine facade
create_from_images is a self-contained image-sequence operation, so it now lives in a focused module while mcp_video.engine keeps the compatibility import. The extraction also reuses the existing quality helper instead of hardcoding preset/CRF arguments.

Constraint: Preserve public imports and tool signatures while continuing low-conflict engine decomposition.

Rejected: Combine this with stabilization or masking | those are output-transforming paths with different FFmpeg risk and should stay separate.

Confidence: high

Scope-risk: narrow

Directive: Keep image-sequence concat path handling covered before changing input normalization or concat file generation.

Tested: ruff check mcp_video/engine.py mcp_video/engine_images.py; ruff format --check mcp_video/engine.py mcp_video/engine_images.py; python3 -m pytest tests/test_engine_advanced.py -k 'create_from_images' -q --tb=short; python3 -m pytest tests/test_cli.py -k 'create_from_images' tests/test_server.py::TestVideoCreateFromImagesTool tests/test_red_team.py -k 'create_from_images' -q --tb=short; python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short

Not-tested: Full slow/real-media suite.
